src/requirements.txt
Django==2.2
djangorestframework==3.9.4
psycopg2==2.8.2
mysqlclient==1.4.2.post1
channels==2.2.0
channels-redis==2.4.0
python-dateutil==2.8.0

uWSGI==2.0.18
gevent==1.4.0
gunicorn==19.9.0
